Added feature toggle VUE_APP_CLAIM_FF
