+++ packages/verification-client/test/Client.spec.js
@@ -20,16 +20,14 @@
  * SOFTWARE.
  */
 
-import Client from '../src/Client'
+import Claim from '../src/Client'
 
-const client = new Client(
+const client = new Claim(
   'wss://ropsten.infura.io/ws/v3/4876e0df8d31475799c8239ba2538c4c',
-    '0xec70947cbb9bbf8b94acaeca861ddbc933b3c789',
-    '0x4bb056574fc19d089e98814d2c8447b2a203b639',
-    '0xd88319a418cf65544f470cacd728b2420e100d20',)
+    '0xec70947cbb9bbf8b94acaeca861ddbc933b3c789')
 
-describe('Client', function () {
-  it('should verify a registered hash', async function (done) {
+  describe('Client', function () {
+   it('should verify a registered hash', async function (done) {
     const res = await client.verifyFile(
       '0x0054f251825dcda879ab6f3dd1e3dd134db01c1a9d1b733775c956b7f179bd0b')
     expect(res).toHaveProperty('issuer')
@@ -41,14 +39,6 @@ describe('Client', function () {
     const res = await client.verifyFile(
       '0xde9b4cf10e72330f5926b26398ba5ffb63b8640407ba30370f21740e16a4484d')
     expect(res.issuer).toBe(null)
-    done()
-  })
-
-  it('should verify a registered hash based on claims only', async function (done) {
-    const res = await client.verifyFile(
-        '0x0be52e65121a2761a837ba5b6702a0961b71b57e4f523739bdca8bdfb026fce5')
-    expect(res).toHaveProperty('issuer')
-    expect(res.issuer).not.toBe(null)
     done()
   })
